backend/requirements.txt
Flask==2.3.3
openai==1.3.9
python-dotenv==1.0.0
Flask-Cors==4.0.0
httpx==0.24.1
qdrant-client==1.6.7